Update simple_state.py
Hello. Congratulations on what you've been able to do with this integration. One of the things needed to use this integration is to be as precise as possible when the tts message actually finishes playing because TTS_END marks when the message is sent to the player and not when it finishes playing. That's why I took this implementation from the StreamAssist integration and looked for the best way to implement it in Hassmic. Practically, when the TTS_END event is executed, the URL of the tts is taken and using mutagen.mp3 to calculate the effective duration of the message. Then in the SimpeState sensor I put a new state tts-finished. There is a problem though with the fact that the wake_word-listening state fires before the message finishes playing and at the end it remains as the last state tts-finished and not wake_word-listening but I don't think that bothers us in ViewAssist automations because both wake_word-listening and tts-finished moments are marked. Please review and let me know if you agree with these implementations.